+++ src/main/java/uk/gov/companieshouse/api/strikeoffobjections/model/entity/ObjectionStatus.java
@@ -1,0 +1,7 @@
+package uk.gov.companieshouse.api.strikeoffobjections.model.entity;
+
+public enum ObjectionStatus {
+    OPEN,
+    PROCESSED,
+    SUBMITTED
+}
