--- src/main/java/uk/gov/companieshouse/api/strikeoffobjections/model/request/RequestStatus.java
@@ -1,7 +1,0 @@
-package uk.gov.companieshouse.api.strikeoffobjections.model.request;
-
-public enum RequestStatus {
-    OPEN,
-    PROCESSED,
-    SUBMITTED
-}
